// src/types/sunStageEnum.ts
export enum SunStageEnum {
    amateurDawn = 'amateurDawn',
    amateurDusk = 'amateurDusk',
    astronomicalDawn = 'astronomicalDawn',
    astronomicalDusk = 'astronomicalDusk',
    blueHourDawnEnd = 'blueHourDawnEnd',
    blueHourDawnStart = 'blueHourDawnStart',
    blueHourDuskEnd = 'blueHourDuskEnd',
    blueHourDuskStart = 'blueHourDuskStart',
    civilDawn = 'civilDawn',
    civilDusk = 'civilDusk',
    goldenHourDawnEnd= 'goldenHourDawnEnd',
    goldenHourDawnStart= 'goldenHourDawnStart',
    goldenHourDuskEnd= 'goldenHourDuskEnd',
    goldenHourDuskStart= 'goldenHourDuskStart',
    nadir= 'nadir',
    nauticalDawn= 'nauticalDawn',
    nauticalDusk= 'nauticalDusk',
    solarNoon= 'solarNoon',
    sunriseEnd= 'sunriseEnd',
    sunriseStart= 'sunriseStart',
    sunsetEnd= 'sunsetEnd',
    sunsetStart= 'sunsetStart'
}